+++ libs/sdk-core/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update `queryRemiScores` method to support pagination and filtering by answer feedback
 - Support new endpoints to download activity log data
 - Updated `feedback` method to allow feedback on specific text blocks
+- Add `entities` property to `FieldMetadata` interface
 
 ### Breaking changes